Update README.md
Purpose

The goal of this project is to gain experience with steganography, cryptanalysis, and random number generators. You will work at the byte level of files to create a bitmap image from scratch. You will programmatically modify such a file, hiding a message inside the image. In the second part, you will break an encryption by finding the key through a weakness in the random number generation.

Objectives

Students will be able to:

    Apply number theory concepts to solve real-world problems.

    Apply algorithms for random number generation, hash functions and encryption.

    Analyze encryption and hash algorithms to find weaknesses and attacks.